added more transparency to waiter image. renamed component files
