+++ fanpage/src/app/components/DinnerSpinnerResult.tsx
