+++ fanpage/src/app/components/Spinner.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { useEffect, useState } from "react";
 import Image from "next/image";
 import WaiterBackground from '../assets/images/waiter-background-spinner.png'
